Handle localized bundle names
For apps that localize their bundle name in InfoPlist.strings (and also remove CFBundleName from App-Info.plist), [[NSBundle mainBundle] infoDictionary] will not contain the key CFBundleName and the app crashes when trying to add the nil bundleName to the iosData dictionary. This fix replaces all direct queries to [[NSBundle mainBundle] infoDictionary] with -[NSBundle objectForInfoDictionaryKey:] as this API will return localized values if available.